+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add configuration `betterGitLineBlame.enableHoverMessages` (true by default) so that users can opt out of hover messages ([#10](https://github.com/mk12/vscode-better-git-line-blame/issues/10)).
+- Show "Commit message loading..." instead of "undefined" when the commit message hasn't loaded yet.
+
 ## 0.2.9
 
 - Add quota system to abort the extension if it gets into a runaway loop.